Increase surrogate/foreign key attributes to BIGINT to accommodate larger data volumes
1. Increase surrogate/foreign key attributes to BIGINT to accommodate
larger data volumes
2. Standardise on use of term "Scale Factor" as per TPC
3.  Optional parameter of no. of data threads. Default to 4 if not
passed.

+++ CREATE_TABLE_DDL/001.call_center.sql
@@ -1,5 +1,5 @@
 CREATE TABLE call_center (
-    cc_call_center_sk integer,
+    cc_call_center_sk bigint,
     cc_call_center_id character(16),
     cc_rec_start_date date,
     cc_rec_end_date text,
