+++ CREATE_TABLE_DDL/016.ship_mode.sql
@@ -1,5 +1,5 @@
 CREATE TABLE ship_mode (
-    sm_ship_mode_sk integer NOT NULL,
+    sm_ship_mode_sk bigint NOT NULL,
     sm_ship_mode_id character(16) NOT NULL,
     sm_type character(30),
     sm_code character(10),
